+++ apps/web/src/app/page.tsx
@@ -18,8 +18,8 @@ export default function Home() {
       </div>
       <div className="absolute inset-x-0 top-1/5 flex flex-col items-center">
         <pre className='px-4 py-2 max-h-[650px] overflow-x-auto rounded-lg border bg-zinc-950 dark:bg-zinc-900 flex justify-between items-center'>
-          <code>npx chai sip -l</code>
-          <CopyButton value='npx chai sip -l' className="ml-4" />
+          <code>npx chaikit sip -l</code>
+          <CopyButton value='npx chaikit sip -l' className="ml-4" />
         </pre>
         {/*  buttons  */}
       </div>