+++ packages/cli/src/index.ts
@@ -11,7 +11,7 @@ async function main() {
   const packageInfo = await getPackageInfo();
 
   const program = new Command()
-    .name("chai")
+    .name("chaikit")
     .description("add components and dependencies to your project")
     .version(
       packageInfo.version || "1.0.0",